requirements.txt
# Core dependencies
python-dotenv
fastapi
uvicorn
pydantic

# Database
psycopg2-binary
sqlalchemy
alembic

# Data processing
pandas
numpy

# Telegram API
telethon
python-telegram-bot

# dbt
dbt-core
dbt-postgres

# Dagster
dagster
dagster-postgres

# YOLO and computer vision
ultralytics
opencv-python
Pillow

# HTTP requests
requests
aiohttp

# Data validation and testing
pytest
pytest-asyncio

# Development tools
black
flake8
mypy

# Logging
loguru

# Configuration
pyyaml 